Whole lot of refactoring and moving classes into packages. Added a new nav item in the drawer menu for my exercises and re-designed the user settings fragment
