+++ app/src/main/java/com/joshrap/liteweight/Database/Daos/ExerciseDao.java
@@ -14,12 +14,16 @@ import java.util.List;
 public interface ExerciseDao {
     @Insert
     void insert(ExerciseEntity entity);
+
     @Update
     void update(ExerciseEntity entity);
+
     @Delete
     void delete(ExerciseEntity entity);
+
     @Query("DELETE FROM exercise_table")
     void deleteAllExercises();
+
     @Query("SELECT * FROM exercise_table")
     List<ExerciseEntity> getAllExercises();
 }
