+++ app/src/main/java/com/joshrap/liteweight/Database/Daos/WorkoutDao.java
@@ -6,6 +6,7 @@ import android.arch.persistence.room.Delete;
 import android.arch.persistence.room.Insert;
 import android.arch.persistence.room.Query;
 import android.arch.persistence.room.Update;
+
 import com.joshrap.liteweight.Database.Entities.*;
 
 import java.util.List;
@@ -14,22 +15,31 @@ import java.util.List;
 public interface WorkoutDao {
     @Insert
     void insert(WorkoutEntity entity);
+
     @Update
     void update(WorkoutEntity entity);
+
     @Delete
     void delete(WorkoutEntity entity);
+
     @Query("DELETE FROM workout_table WHERE workout=:name")
     void deleteEntireWorkout(String name);
+
     @Query("DELETE FROM workout_table")
     void deleteAllWorkouts();
+
     @Query("DELETE FROM workout_table WHERE exercise=:exerciseName AND workout=:workoutName AND day=:day")
     void deleteSpecificExerciseFromWorkout(String workoutName, String exerciseName, int day);
+
     @Query("DELETE FROM workout_table WHERE exercise=:exerciseName")
     void deleteExerciseFromWorkouts(String exerciseName);
+
     @Query("SELECT * FROM workout_table")
     LiveData<List<WorkoutEntity>> getAllWorkouts();
+
     @Query("SELECT * FROM workout_table WHERE workout=:workoutName")
     List<WorkoutEntity> getExercises(String workoutName);
+
     @Query("UPDATE workout_table SET exercise=:newName WHERE exercise=:oldName")
     void updateExerciseName(String oldName, String newName);
 }
