+++ app/src/main/java/com/joshrap/liteweight/Fragments/AboutFragment.java
@@ -7,7 +7,9 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
 import com.joshrap.liteweight.*;
+import com.joshrap.liteweight.Globals.Variables;
 
 public class AboutFragment extends Fragment {
     private View view;
