+++ app/src/main/java/com/joshrap/liteweight/Helpers/WorkoutHelper.java
@@ -1,0 +1,9 @@
+package com.joshrap.liteweight.Helpers;
+
+public class WorkoutHelper {
+    public static String generateDayTitle(int num, int numDays) {
+        int weekNum = (num / numDays) + 1;
+        int dayNum = (num % numDays) + 1;
+        return "W" + weekNum + ":D" + dayNum;
+    }
+}
