+++ app/src/main/res/drawable/list_activated_background.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_activated="true" android:drawable="@color/row_activated" />
-    <item android:state_checked="true" android:drawable="@color/row_activated" />
-    <item android:state_pressed="true" android:drawable="@color/row_pressed" />
+    <item android:drawable="@color/row_activated" android:state_activated="true" />
+    <item android:drawable="@color/row_activated" android:state_checked="true" />
+    <item android:drawable="@color/row_pressed" android:state_pressed="true" />
     <item android:drawable="@android:color/transparent" />
 </selector>
