+++ app/src/main/res/drawable/my_exercises_icon.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="100dp"
+    android:height="1250dp"
+    android:tint="#7E0615"
+    android:viewportWidth="100"
+    android:viewportHeight="125">
+    <path
+        android:fillColor="#00000000"
+        android:pathData="M95,94.5L5,94.5"
+        android:strokeWidth="1"
+        android:strokeColor="#000000" />
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M16.392,5c-0.55,0 -1,0.45 -1,1v14c0,0.55 0.45,1 1,1c0.55,0 1,-0.45 1,-1V6C17.392,5.45 16.942,5 16.392,5zM13.392,8c-0.55,0 -1,0.45 -1,1v8c0,0.55 0.45,1 1,1c0.55,0 1,-0.45 1,-1V9C14.392,8.45 13.942,8 13.392,8zM10.392,10c-0.55,0 -1,0.45 -1,1v4c0,0.55 0.45,1 1,1c0.55,0 1,-0.45 1,-1v-4C11.392,10.45 10.942,10 10.392,10zM19.392,5c-0.55,0 -1,0.45 -1,1v14c0,0.55 0.45,1 1,1c0.55,0 1,-0.45 1,-1V6C20.392,5.45 19.942,5 19.392,5zM84.392,5c-0.55,0 -1,0.45 -1,1v14c0,0.55 0.45,1 1,1c0.55,0 1,-0.45 1,-1V6C85.392,5.45 84.942,5 84.392,5zM81.392,5c-0.55,0 -1,0.45 -1,1v14c0,0.55 0.45,1 1,1c0.55,0 1,-0.45 1,-1V6C82.392,5.45 81.942,5 81.392,5zM87.392,8c-0.55,0 -1,0.45 -1,1v8c0,0.55 0.45,1 1,1c0.55,0 1,-0.45 1,-1V9C88.392,8.45 87.942,8 87.392,8zM90.392,10c-0.55,0 -1,0.45 -1,1v4c0,0.55 0.45,1 1,1c0.55,0 1,-0.45 1,-1v-4C91.392,10.45 90.942,10 90.392,10zM58.493,11.024h-1.151c-0.542,0 -0.982,0.438 -0.995,0.976H43.653c-0.013,-0.539 -0.454,-0.976 -0.995,-0.976h-1.151c-0.542,0 -0.982,0.438 -0.995,0.976H21.001v1h58v-1H59.488C59.475,11.461 59.034,11.024 58.493,11.024zM54,30.992c0,-3.037 -1.028,-5.016 -4.001,-5.016c-3.016,0 -3.999,1.978 -3.999,5.016c0,3.283 2,5.984 3.999,5.984C52.031,36.976 54,34.275 54,30.992zM56.315,14l-0.137,10.17c0.001,0.053 0.001,0.1 0.002,0.157c0.02,1.064 0.838,1.65 0.838,1.65s-0.425,-0.041 -0.861,-0.282l-0.153,11.31h-0.014c0,0.55 -0.446,0.995 -0.995,0.995h-9.991c-0.55,0 -0.995,-0.446 -0.995,-0.995h-0.014l-0.153,-11.314c-0.438,0.245 -0.868,0.285 -0.868,0.285s0.819,-0.586 0.838,-1.65c0.003,-0.161 0.005,-0.303 0.006,-0.436L43.685,14H40.47l-0.463,24.809c0,1.997 1.428,4.124 1.994,4.178c0,0.002 -0.001,0.008 -0.001,0.01l0.001,0.01h0.296c1.688,0 1.719,1.993 1.719,1.993s-0.363,-0.997 -1.716,-0.997c-0.059,0 -0.109,0 -0.164,0l1.707,12.612c0.189,0.938 1.147,1.357 1.147,1.357s-0.47,0.066 -0.985,-0.027c-0.001,0.006 -0.004,0.011 -0.004,0.017l-2,17.865c0,0.049 0.007,0.084 0.015,0.118c-0.002,0.025 -0.015,0.047 -0.015,0.073c0,0.02 0.01,0.036 0.011,0.055l0.964,16.909c0,0.55 0.45,1 1,1h1.729c0.55,0 1,-0.45 1,-1l0.77,-15.93C47.449,76.074 45.989,75 45.989,75s0.748,0.188 1.415,0.643c0.026,-0.174 0.051,-0.387 0.075,-0.653l2.375,-11.013h0.275l2.375,11.013c0.024,0.268 0.049,0.481 0.076,0.656C53.239,75.195 53.97,75 53.97,75s-1.434,1.073 -1.461,2.053l0.77,15.93c0,0.55 0.45,1 1,1h1.729c0.55,0 1,-0.45 1,-1l0.964,-16.909c0.001,-0.019 0.011,-0.036 0.011,-0.055c0,-0.026 -0.013,-0.047 -0.015,-0.073c0.007,-0.034 0.015,-0.069 0.015,-0.118l-2,-17.865c0,-0.016 -0.006,-0.03 -0.009,-0.044c-0.513,0.101 -0.977,0.054 -0.977,0.054s0.973,-0.421 1.16,-1.364l1.706,-12.606c-0.055,0 -0.105,0 -0.164,0c-1.352,0 -1.716,0.997 -1.716,0.997s0.031,-1.993 1.719,-1.993h0.296L58,42.997c0,-0.002 -0.001,-0.008 -0.001,-0.01c0.566,-0.054 1.994,-2.181 1.994,-4.178L59.53,14H56.315z" />
+</vector>
