+++ app/src/main/res/drawable/video_icon.xml
@@ -1,5 +1,10 @@
-<vector android:height="50dp" android:tint="#7E0615"
-    android:viewportHeight="24.0" android:viewportWidth="24.0"
-    android:width="50dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M8,5v14l11,-7z"/>
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="50dp"
+    android:height="50dp"
+    android:tint="#7E0615"
+    android:viewportWidth="24.0"
+    android:viewportHeight="24.0">
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M8,5v14l11,-7z" />
 </vector>
