+++ app/src/main/res/layout/popup_add_exercise.xml
@@ -34,7 +34,6 @@
         android:layout_marginEnd="16dp"
         android:layout_marginRight="16dp"
         android:layout_marginBottom="8dp"
-        android:entries="@array/exercise_focuses"
         android:spinnerMode="dropdown"
         app:layout_constraintBottom_toTopOf="@+id/scroll_view"
         app:layout_constraintEnd_toEndOf="parent"
